Transforming mapmaking.py into a module
